test: Cover PSR-16 implementation with unit tests

+++ test/Integration/Sql/Base.php
@@ -1,6 +1,7 @@
 <?php
+
 /**
- * Copyright 2016-2021 Horde LLC (http://www.horde.org/)
+ * Copyright 2016-2026 Horde LLC (http://www.horde.org/)
  *
  * See the enclosed file LICENSE for license information (LGPL). If you
  * did not receive this file, see http://www.horde.org/licenses/lgpl21.
@@ -10,10 +11,10 @@
  * @license  http://www.horde.org/licenses/lgpl21 LGPL 2.1
  * @package  Cache
  */
-namespace Horde\Cache\Test\Sql;
+
+namespace Horde\Cache\Test\Integration\Sql;
+
 use Horde\Cache\Test\TestBase;
-use Horde_Log_Logger;
-use Horde_Log_Handler_Cli;
 use PEAR_Config;
 use Horde_Db_Migration_Migrator;
 use Horde\Cache\Cache;
@@ -29,11 +30,12 @@ use Horde\Cache\SqlStorage;
  */
 class Base extends TestBase
 {
-    protected function _getCache($params = array())
+    protected $db;
+    protected $migrator;
+
+    protected function _getCache($params = [])
     {
-        $logger = new Horde_Log_Logger(new Horde_Log_Handler_Cli());
-        $this->db->setLogger($logger);
-        $dir = __DIR__ . '/../../migration/Horde/Cache';
+        $dir = __DIR__ . '/../../../migration/Horde/Cache';
         if (!is_dir($dir)) {
             error_reporting(E_ALL & ~E_DEPRECATED);
             $dir = PEAR_Config::singleton()
@@ -41,20 +43,26 @@ class Base extends TestBase
                 . '/Horde_Cache/migration';
             error_reporting(E_ALL | E_STRICT);
         }
-        
+
         if (class_exists('Horde_Db_Migration_Migrator')) {
             $this->migrator = new Horde_Db_Migration_Migrator(
-                        $this->db,
-                        $logger,
-                        array('migrationsPath' => $dir,
-                            'schemaTableName' => 'horde_cache_schema_info'));
-                    $this->migrator->up();
-       
+                $this->db,
+                null,
+                ['migrationsPath' => $dir,
+                    'schemaTableName' => 'horde_cache_schema_info']
+            );
+            $this->migrator->up();
+
+            $merged = array_merge(
+                ['db' => $this->db],
+                $params
+            );
+
             return new Cache(
-                new SqlStorage(array_merge(
-                    ['db'   => $this->db],
-                    $params
-                ))
+                new SqlStorage(
+                    db: $merged['db'],
+                    table: $merged['table'] ?? 'horde_cache'
+                )
             );
         }
     }
@@ -63,7 +71,9 @@ class Base extends TestBase
     public function tearDown(): void
     {
         parent::tearDown();
-        $this->db->delete('DELETE FROM horde_cache');
+        if ($this->db) {
+            $this->db->delete('DELETE FROM horde_cache');
+        }
         if ($this->migrator) {
             $this->migrator->down();
         }
